tidy up legacy token handling
