+++ .env.example
@@ -5,7 +5,7 @@
 #
 # Uncomment only if you need to override defaults:
 
-# Cognito flow is enabled by default (true). Set false to use legacy JWT auth.
+# Cognito flow comes from SSM (e2e-cognito-flow-enabled). Set false to use legacy JWT auth.
 # export CYPRESS_COGNITO_FLOW_ENABLED=false
 
 # Legacy JWT auth — only when CYPRESS_COGNITO_FLOW_ENABLED=false
