+++ cypress/plugins/environment-config.js
@@ -1,15 +1,5 @@
 const { fetchCognitoToken, cognitoFlowEnabled } = require("./cognito-helper");
-
-const LEGACY_TOKEN_KEYS = {
-  development: "E2E_ACCESS_TOKEN_DEVELOPMENT",
-  staging: "E2E_ACCESS_TOKEN_STAGING",
-  production: "E2E_ACCESS_TOKEN_PRODUCTION",
-};
-
-function resolveLegacyToken(configEnv, environment) {
-  const key = LEGACY_TOKEN_KEYS[environment];
-  return key ? configEnv[key] : undefined;
-}
+const { resolveLegacyToken } = require("./legacy-auth");
 
 const setEnvironmentConfig = async (on, config) => {
   // setting page paths
@@ -51,6 +41,4 @@ const setEnvironmentConfig = async (on, config) => {
 
 module.exports = {
   setEnvironmentConfig,
-  resolveLegacyToken,
-  LEGACY_TOKEN_KEYS,
 };
